Second sub-command: generate_films_index
This command takes the entries from base_index.tsv and carries out
an OMDB search to gather their metadata. If the OMDB search fails,
or if the user prefers not to supply an OMDB API-key, the command
falls back on an IMDB scrape-search, which is considerably slower.
The output is a 6-column TSV, with the following columns: filmname,
filmyear, director, genres, actors, filmpath. This file will be
used by the create* commands to generate folder hierarchies.