Subcommands 4-6: generate_actors_list, generate_actors_filmography, populate_actors_tree
The first of these commands creates a list of all oscar-nominated actors
and actresses, along with their IMDB namecodes. The second command uses
the namecodes to retrieve and store filmography information about the
actors, namely a list of their high-rated films and the years they were
released. The third command uses this filmography data along with the
base_index.tsv file to populate a "Films by Actor" folder tree. This
catches all good films in which an actor was involved, even if he or she
was not top-billed for that film.