Adding the nff helper subcommand, see details below
nff stands for normalize_film_files and is meant to convert bare
filmnames into the format '(year) filmname' which is the default
format expected by the generate_base_index function.

You can supply a custom regex on the command-line if you wish to
normalize to some other naming-format. By default, this function
tries to find a film-year within the filename itself and reformats
the filename accordingly; if this fails, it tries to look up the
non-standard filmname on IMDB and scrapes & adds the year-info if
an exact IMDB match is found. It prints out a warning message if
no unique match could be found.

You can also specify the -i or --interactive flag which causes the
program to display an input prompt for all filenames which fail the
IMDB look-up, so that the user can enter the correct name directly.